requirements.txt
# Data
yfinance==0.2.36
pandas==2.2.1
numpy==1.26.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.10

# NLP
transformers==4.40.0
torch==2.2.2
sentencepiece==0.2.0

# Modeling
xgboost==2.0.3
pytorch-forecasting==1.0.0
pytorch-lightning==2.2.3
scikit-learn==1.4.2

# MLOps
mlflow==2.12.1
python-dotenv==1.0.1

# News
requests==2.31.0
newsapi-python==0.2.7

# Dashboard
streamlit==1.33.0
plotly==5.20.0

# Scheduling
apache-airflow==2.9.0

streamlit==1.33.0
plotly==5.20.0
pandas==2.2.1
numpy==1.26.4
torch==2.2.2
transformers==4.40.0
xgboost==2.0.3
scikit-learn==1.4.2
mlflow==2.12.1
python-dotenv==1.0.1
```
